Make internal links HTTPS ready
Internal links shouldn't hardcode http, but rather go over the protocol the website
is currently surfed with, or if HSTS is enabled for the internal site, https
should be hardcoded.
Therefore, drop the {http,https}
prefix from minetest-internal (that is any subdomain of minetest.net or the domain
itself) links, except forum.minetest.net, where all occurences are changed to https.