[WIP] Channel description improvements 2.
Making all channels match code style and text formatting.
Adding a more direct and clear description to each channel.
Making use of uppercase for aesthetics and to make things more relevant.
Adding webchat for Minetest-Hub. In the topic of the channel we have a link for the channels website and it also says about voice and asking an op for it.
Adding webchat for Minetest-Hub. Mainly to match style.

Opinions needed.