Add a forum redirection page
This is done in the 404 error page as it will match any invalid URL,
including `/forum/viewtopic.php`. We can't create a page with a custom
permalink set to `/forum/viewtopic.php`, as GitHub Pages doesn't set
the MIME type correctly, which causes the browser to download the page.

This closes #187.